#4 refactor: 생성자 호출을 정적 팩토리 메소드 호출로 변경
