+++ backend/src/main/java/codesquard/app/api/oauth/request/OauthLoginRequest.java
@@ -1,7 +1,12 @@
-package codesquard.app.api.oauth;
+package codesquard.app.api.oauth.request;
 
 import javax.validation.constraints.Pattern;
 
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@ToString
 public class OauthLoginRequest {
 	@Pattern(regexp = "^[가-힣a-zA-Z0-9]{2,12}$", message = "아이디는 띄어쓰기 없이 한글, 영문, 숫자로 구성되며 2~12글자로 구성되어야 합니다.")
 	private String loginId;
