+++ backend/src/main/java/codesquard/app/api/oauth/response/OauthUserProfileResponse.java
@@ -1,12 +1,18 @@
 package codesquard.app.api.oauth.response;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 @Getter
 @ToString
-@RequiredArgsConstructor
 public class OauthUserProfileResponse {
 	private final String socialLoginId;
+
+	private OauthUserProfileResponse(String socialLoginId) {
+		this.socialLoginId = socialLoginId;
+	}
+
+	public static OauthUserProfileResponse create(String socialLoginId) {
+		return new OauthUserProfileResponse(socialLoginId);
+	}
 }
